test=develop alter test_rpn_target_assign_op.py
